+++ src/main/java/com/djarjo/text/TextHelper.java
@@ -51,7 +51,8 @@ public class TextHelper {
 					"&=!%{}" +
 					"()[]/";
 	/**
-	 * Replaces tabs with space and condenses multiple spaces into one single space .<br />
+	 * Replaces tabs with space and condenses multiple spaces into one single space
+	 * .<br />
 	 * Value for parameter 'option' in {@link #condense(String, int)}.<br /> Multiple
 	 * options can be added.
 	 */
@@ -74,7 +75,8 @@ public class TextHelper {
 	/**
 	 * Removes end of line comments starting with "//..." until end of line <br /> Value
 	 * for
-	 * parameter 'option' in {@link #condense(String, int)}.<br /> Multiple options can be
+	 * parameter 'option' in {@link #condense(String, int)}.<br /> Multiple options
+	 * can be
 	 * added.
 	 */
 	public final static int REMOVE_COMMENT_EOL = 16;
@@ -95,7 +97,8 @@ public class TextHelper {
 	public final static int REMOVE_COMMENT_LONG = 128;
 	/**
 	 * Removes remark lines starting with "# " or "-- "until end of line<br /> Value for
-	 * parameter 'option' in {@link #condense(String, int)}.<br /> Multiple options can be
+	 * parameter 'option' in {@link #condense(String, int)}.<br /> Multiple options
+	 * can be
 	 * added.
 	 */
 	public final static int REMOVE_COMMENT_LINES = 256;
@@ -128,7 +131,8 @@ public class TextHelper {
 	public final static int REMOVE_COMMENTS = REMOVE_COMMENT_EOL
 			+ REMOVE_COMMENT_LONG + REMOVE_COMMENT_LINES + REMOVE_COMMENT_XML;
 	/**
-	 * Removes all comments, blanks and empty lines.<br /> Value for parameter 'option' in
+	 * Removes all comments, blanks and empty lines.<br /> Value for parameter
+	 * 'option' in
 	 * {@link #condense(String, int)}
 	 */
 	public final static int REMOVE_ALL = 8191;
@@ -212,7 +216,8 @@ public class TextHelper {
 	 * space and multiple spaces with a single one. Spaces at start and end are removed.
 	 *
 	 * @param str well ... the string
-	 * @return Returns condensed string or {@code null} if the given string is {@code null}
+	 * @return Returns condensed string or {@code null} if the given string is {@code
+	 * null}
 	 */
 	public static String cleanupString( String str ) {
 		if ( str == null ) {
@@ -230,7 +235,8 @@ public class TextHelper {
 	 * space and multiple spaces with a single one. Spaces at start and end are removed.
 	 *
 	 * @param buf buffer to be cleaned
-	 * @return Returns condensed string or {@code null} if the given string is {@code null}
+	 * @return Returns condensed string or {@code null} if the given string is {@code
+	 * null}
 	 */
 	public static String cleanupString( StringBuilder buf ) {
 		if ( buf == null ) {
@@ -476,7 +482,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Condenses the given text according to the given options. A line with a backslash '\'
+	 * Condenses the given text according to the given options. A line with a backslash
+	 * '\'
 	 * at its end will be continued in the next line. {@code options} is the sum of:
 	 * <table>
 	 * <caption>Options</caption>
@@ -899,12 +906,14 @@ public class TextHelper {
 	 * Returns a new string of contiguous characters out of the given string starting at
 	 * position <code>offset</code>.
 	 * <p>
-	 * The sequence of contiguous characters starts with the first non-blank character . It
+	 * The sequence of contiguous characters starts with the first non-blank character
+	 * . It
 	 * ends when any character is found which is neither a letter, a digit nor an
 	 * underscore.
 	 * </p>
 	 * <p>
-	 * If special characters should be included in the string (like a dot '.' in a package
+	 * If special characters should be included in the string (like a dot '.' in a
+	 * package
 	 * name) these have to be given in the parameter
 	 * <code>includes</code>.
 	 * </p>
@@ -1051,7 +1060,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Gets the word from the given string starting at the given offset. A word consists of
+	 * Gets the word from the given string starting at the given offset. A word
+	 * consists of
 	 * letters. It finishes when anything comes in the text which is not a letter.
 	 *
 	 * @param str The text from which to extract the word
@@ -1100,7 +1110,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Checks the given string if it means <em>false</em>. This is the case when the string
+	 * Checks the given string if it means <em>false</em>. This is the case when the
+	 * string
 	 * is "false", "no" or "0". Everything else returns
 	 * <em>false</em>. Matching is not case sensitive.
 	 *
@@ -1115,14 +1126,16 @@ public class TextHelper {
 	}
 
 	/**
-	 * Computes the <em>Levenshtein</em> distance between the two words. If the distance is
+	 * Computes the <em>Levenshtein</em> distance between the two words. If the
+	 * distance is
 	 * smaller than 20% (rounded up) of the length of the first word, then {@code true}
 	 * will
 	 * be returned.
 	 *
 	 * @param word1 The first word
 	 * @param word2 The second word
-	 * @return Returns {@code false} if more than 20% of the characters must be changed to
+	 * @return Returns {@code false} if more than 20% of the characters must be
+	 * changed to
 	 * get the second word
 	 */
 	public static boolean isSameByLevenshtein( String word1, String word2 ) {
@@ -1147,7 +1160,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Makes a string from the given object. If the object is a collection (array or list)
+	 * Makes a string from the given object. If the object is a collection (array or
+	 * list)
 	 * its values will be returned separated by commas.
 	 *
 	 * @param object some object
@@ -1374,7 +1388,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Parses the given text as a signed long integer value (64 bit). Skips leading blanks.
+	 * Parses the given text as a signed long integer value (64 bit). Skips leading
+	 * blanks.
 	 * Reads digits for the value until the first non-digit.
 	 *
 	 * @param text Text to be parsed.
@@ -1542,7 +1557,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Parses the given text as a signed short value (16 bit). Skips leading blanks. Reads
+	 * Parses the given text as a signed short value (16 bit). Skips leading blanks.
+	 * Reads
 	 * digits for the value until the first non-digit.
 	 *
 	 * @param text Text to be parsed.
@@ -1566,30 +1582,33 @@ public class TextHelper {
 	}
 
 	/**
-	 * Parses {@code text} into {@code TimeZone}
+	 * Parses {@code text} into a valid timezone.
+	 * <p>
+	 * If text is not empty but cannot be parsed then system default timezone will be
+	 * returned.
+	 * </p>
 	 *
 	 * @param text ... to parse
-	 * @return {@code TimeZone} or {@code null} if {@code text} was null or empty
+	 * @return valid timeZone or {@code null} if {@code text} was null or empty
 	 */
-	public static TimeZone parseTimeZone( final String text ) {
+	public static ZoneId parseTimeZone( final String text ) {
 		if ( isEmpty( text ) ) return null;
 
-		//--- its a Duration
+		//--- it's a Duration
 		if ( text.startsWith( "PT" ) || text.startsWith( "-PT" ) ) {
 			Duration duration = Duration.parse( text );
-			return BaseConverter.toTimeZoneFromMinutes( duration.toMinutes() );
+			return BaseConverter.toZoneIdFromMinutes( (int) duration.toMinutes() );
 		}
 
 		//--- if its just an integer in minutes then it must be a multiple of 60 or 90
 		Integer minutes = parseInteger( text );
 		if ( (minutes != null) && ((minutes % 60 == 0) || (minutes % 90 == 0)) ) {
-			return BaseConverter.toTimeZoneFromMinutes( minutes );
+			return BaseConverter.toZoneIdFromMinutes( minutes );
 		}
 
 		//--- seems to be a ZoneId
 		try {
-			ZoneId zoneId = ZoneId.of( text );
-			return TimeZone.getTimeZone( zoneId );
+			return ZoneId.of( text );
 		} catch ( DateTimeException e ) {
 			return null;
 		}
@@ -1771,7 +1790,8 @@ public class TextHelper {
 	}
 
 	/**
-	 * Returns a new string from which leading and trailing blanks [ \n\r\t] are removed .
+	 * Returns a new string from which leading and trailing blanks [ \n\r\t] are
+	 * removed .
 	 * If the given string is {@code null} or empty then {@code null} will be returned
 	 *
 	 * @param str String to trim